move scrolling focus down to 2/3 of the screen (#442)
* move scrolling focus to 2/3 of the screen

* scroll to bottom only if next row is not already visible
